requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
dateparser==1.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pycryptodome==3.20.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
six==1.16.0
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
